Improve Item List Page: Add Informative Text and Modal Functionality
Updated the "Add Item" page to provide more informative text, explaining that the list of items is displayed and users can click the "Create" button to add new data.
Styled the heading and description to be more visually prominent and clear.
Implemented functionality to show a modal when the "Create" button is clicked, enhancing user experience for item management.